tests/wpt/tests/css/css-text/parsing/text-wrap-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text: getComputedStyle().textWrap</title>
<link rel="author" title="Tim Nguyen" href="https://github.com/nt1m">
<link rel="help" href="https://drafts.csswg.org/css-text-4/#text-wrap">
<meta name="assert" content="text-wrap computed value is '<text-wrap-mode> || <text-wrap-style>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("text-wrap", "wrap");
test_computed_value("text-wrap", "nowrap");

test_computed_value("text-wrap", "auto", "wrap");
test_computed_value("text-wrap", "balance");
test_computed_value("text-wrap", "stable");
test_computed_value("text-wrap", "pretty");

test_computed_value("text-wrap", "wrap auto", "wrap");
test_computed_value("text-wrap", "wrap balance", "balance");
test_computed_value("text-wrap", "wrap pretty", "pretty");
test_computed_value("text-wrap", "wrap stable", "stable");
test_computed_value("text-wrap", "auto wrap", "wrap");
test_computed_value("text-wrap", "balance wrap", "balance");
test_computed_value("text-wrap", "pretty wrap", "pretty");
test_computed_value("text-wrap", "stable wrap", "stable");

test_computed_value("text-wrap", "nowrap auto", "nowrap");
test_computed_value("text-wrap", "nowrap balance");
test_computed_value("text-wrap", "nowrap pretty");
test_computed_value("text-wrap", "nowrap stable");
test_computed_value("text-wrap", "auto nowrap", "nowrap");
test_computed_value("text-wrap", "balance nowrap", "nowrap balance");
test_computed_value("text-wrap", "pretty nowrap", "nowrap pretty");
test_computed_value("text-wrap", "stable nowrap", "nowrap stable");
</script>
</body>
</html>
